add check map

+++ get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: hmouis <hmouis@1337.ma>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/14 09:38:28 by hmouis            #+#    #+#             */
-/*   Updated: 2025/02/14 09:38:30 by hmouis           ###   ########.fr       */
+/*   Updated: 2025/02/14 14:00:54 by hmouis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,9 +29,9 @@ char	*ft_strchr(char *s, char c)
 	return (NULL);
 }
 
-size_t	ft_strlen(char *s)
+int	ft_strlen(char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (!s)
